Pulled out scheduler into own module
